updated challenger sdk to latest
