+++ src/util/challengerSdk/challenger.pda.ts
@@ -1,36 +1,36 @@
 import { PublicKey } from '@solana/web3.js';
-import { CHALLENGER_PROG_ID } from '../index';
+import { CHALLENGER_PROG_ID } from './index';
 
 export const findCruxAuthorityPDA = async (crux: PublicKey) => {
-    return PublicKey.findProgramAddressSync(
+    return PublicKey.findProgramAddress(
         [crux.toBytes()],
         CHALLENGER_PROG_ID
     );
 };
 
 export const findCruxTreasuryPDA = async (crux: PublicKey) => {
-    return PublicKey.findProgramAddressSync(
+    return PublicKey.findProgramAddress(
         [Buffer.from('treasury'), crux.toBytes()],
         CHALLENGER_PROG_ID
     );
 };
 
 export const findUserProfilePDA = async (crux: PublicKey, profileOwner: PublicKey) => {
-    return PublicKey.findProgramAddressSync(
+    return PublicKey.findProgramAddress(
         [Buffer.from('user_profile'), crux.toBytes(), profileOwner.toBytes()],
         CHALLENGER_PROG_ID
     );
 };
 
 export const findChallengePDA = async (crux: PublicKey, challengeSeed: PublicKey) => {
-    return PublicKey.findProgramAddressSync(
+    return PublicKey.findProgramAddress(
         [Buffer.from('challenge'), crux.toBytes(), challengeSeed.toBytes()],
         CHALLENGER_PROG_ID
     );
 };
 
 export const findSubmissionPDA = async (challenge: PublicKey, userProfile: PublicKey, ) => {
-    return PublicKey.findProgramAddressSync(
+    return PublicKey.findProgramAddress(
         [Buffer.from('submission'), challenge.toBytes(), userProfile.toBytes()],
         CHALLENGER_PROG_ID
     );
