--- src/util/challengerSdk/challenger/index.ts
@@ -1,2 +1,0 @@
-export * from './challenger.client';
-export * from './challenger.pda';
