+++ src/util/challengerSdk/index.ts
@@ -1,6 +1,8 @@
 import { PublicKey } from '@solana/web3.js';
 
-export * from './challenger';
+export * from './challenger.client';
+export * from './challenger.pda';
+export * from './challenger.types'
 
 export const CHALLENGER_PROG_ID = new PublicKey(
     'CRuXQ86F4m6VfRHa7VACNbQKJoSioG3gcpui9BH2YNWa'
